initial files

+++ models/listing.js
@@ -12,6 +12,7 @@ const listingSchema = new Schema({
   image: {
     url: {
       type: String,
+      default: "",   
     },
     filename: {
       type: String,
@@ -26,17 +27,22 @@ const listingSchema = new Schema({
   reviews: [
     {
       type: Schema.Types.ObjectId,
-      ref: "Review", 
+      ref: "Review",
     },
   ],
-});
 
-listingSchema.post("findOneAndDelete", async (listing) => {
-  if(listing) {
-     await Review.deleteMany({ _id: { $in: listing.reviews } });
+  owner: {
+    type: Schema.Types.ObjectId,
+    ref: "User",
   }
 });
 
+
+listingSchema.post("findOneAndDelete", async (listing) => {
+  if (listing && listing.reviews.length > 0) {
+    await Review.deleteMany({ _id: { $in: listing.reviews } });
+  }
+});
 
 const Listing = mongoose.model("Listing", listingSchema);
 module.exports = Listing;
